+++ inst/install-github.R
@@ -756,7 +756,7 @@ function(...) {
     dependencies <- intersect(dependencies, names(pkg))
   
     #remove standard dependencies
-    setdiff(dependencies, tolower(standardise_dep(TRUE)))
+    setdiff(dependencies, tolower(standardise_dep(c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances"))))
   }
   
   #' @export