+++ NEWS.md
@@ -4,6 +4,8 @@
 * `install_github()` now uses credentials from the git credential store,
   if `GITHUB_PAT` and `GITHUB_TOKEN` are not set.
 * Remotes field accepts explicit package names: `<pkgname>=<type>::<username>/<repo>` (#719, @heavywatal).
+* Add `remote_precedence` parameter to `dev_package_deps()` which allows user to choose whether 
+  remote sources should have a priority over cran sources. (#732, @maksymiuks)
 
 # remotes 2.4.2