+++ NEWS.md
@@ -7,8 +7,8 @@
 * Add `remote_precedence` parameter to `dev_package_deps()` which allows user to choose whether 
   remote sources should have a priority over cran sources. (#732, @maksymiuks)
 
-* Fix `get_extra_deps()` helper not excluding Enhances dependencies what
-  could lead to errors if developer specified version requirement.
+* `dev_package_deps()` now works for packages with `Enhances` dependencies
+  (#711, @maksymiuks).
 
 # remotes 2.4.2